refactor: move ReactNativeLauncher to root
