+++ src/screens/connectors/LauncherView.js
@@ -5,7 +5,7 @@ import {StyleSheet, View} from 'react-native'
 import templateConnector from '../../../connectors/template/dist/webviewScript'
 import sncfConnector from '../../../connectors/sncf/dist/webviewScript'
 import blablacarConnector from '../../../connectors/blablacar/dist/webviewScript'
-import ReactNativeLauncher from './libs/ReactNativeLauncher'
+import ReactNativeLauncher from '../../libs/ReactNativeLauncher'
 import CookieManager from '@react-native-cookies/cookies'
 import debounce from 'lodash/debounce'
 import {withClient} from 'cozy-client'
